Update : restructured the context and put them in a different folders
