+++ backend/package.json
@@ -29,7 +29,8 @@
     "kareem": "^2.6.3",
     "mongoose": "^8.5.1",
     "prettier": "^3.3.3",
-    "resend": "^4.0.0"
+    "resend": "^4.0.0",
+    "socket.io": "^4.7.5"
   },
   "devDependencies": {
     "@eslint/js": "^9.6.0",
@@ -45,6 +46,8 @@
     "globals": "^15.7.0",
     "jest": "^29.7.0",
     "nodemon": "^3.1.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "supertest": "^7.0.0",
     "ts-jest": "^29.2.3",
     "ts-node": "^10.9.2",
