+++ frontend/package.json
@@ -23,7 +23,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.52.2",
-    "react-router-dom": "^6.26.0"
+    "react-router-dom": "^6.26.0",
+    "socket.io-client": "^4.7.5"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.7",
