+++ frontend/src/Auth/SignUp.tsx
@@ -9,7 +9,7 @@ import SignUpUserApi from "../packages/Api/UserApi/SignUpUserApi"
 import { AnimatePresence, motion } from "framer-motion"
 import { Loading } from "../components/Containers/Loading"
 import "../components/Containers/ContainersStyles.css"
-import { useUser } from "../Context/useUser"
+import { useUser } from "../Context/User/useUser"
 
 
 export const SignUp: React.FC<SignUpModalProps> = ({ onSignUpSuccessful }) => {
@@ -75,7 +75,7 @@ export const SignUp: React.FC<SignUpModalProps> = ({ onSignUpSuccessful }) => {
     // handles the bg color
     useEffect(() => {
       document.body.style.background = "rgb(0,14,36)";
-      document.body.style.background = "linear-gradient(90deg, rgba(0,14,36,1) 4%, rgba(166,17,180,0.013064600840336116) 100%, rgba(0,212,255,1) 100%)";
+      document.body.style.background = "linear-gradient(90deg, rgba(0,14,36,1) 4%)";
     }, []);
