+++ frontend/src/Auth/UpdatePassword.tsx
@@ -78,7 +78,7 @@ export const UpdatePassword: React.FC = () => {
   // handles the bg color
   useEffect(() => {
     document.body.style.background = "rgb(0,14,36)";
-    document.body.style.background = "linear-gradient(90deg, rgba(0,14,36,1) 4%, rgba(166,17,180,0.013064600840336116) 100%, rgba(0,212,255,1) 100%)";
+    document.body.style.background = "linear-gradient(90deg, rgba(0,14,36,1) 4%)";
   }, []);
