+++ frontend/src/components/Containers/AddNewTaskContainer.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from '../../Context/UseTheme';
+import { useTheme } from '../../Context/Theme/UseTheme';
 import React from 'react';
 import {
   onCloseContainerProp,
@@ -10,7 +10,7 @@ import { useForm, Controller } from 'react-hook-form';
 import CreateTaskApi from '../../packages/Api/BoardApi/CreateTaskApi';
 import { SubTaskInputContainer } from './SubTaskColumnContainer';
 import { TaskStatusDropdown } from './TaskStatusContainer';
-import { useUser } from '../../Context/useUser';
+import { useUser } from '../../Context/User/useUser';
 import { openCustomNotification } from '../../utils/notificationUtil';
 import { NotificationContainerStyle } from '../../utils/NotificationContainerStyle';
 import { SuccessIcon } from '../../Icons/SuccessIcon';
