+++ frontend/src/components/Containers/EditBoardContainer.tsx
@@ -2,10 +2,10 @@ import React, { useState } from 'react';
 import { PrimaryBtnSmall } from '../Buttons/PrimaryBtnSmall';
 import { EditProps } from '../Interface/EditBoardInterface';
 import './ContainersStyles.css';
-import { useTheme } from '../../Context/UseTheme';
+import { useTheme } from '../../Context/Theme/UseTheme';
 import { BoardColumnContainer } from './BoardColumnContainer';
 import { useForm } from 'react-hook-form';
-import { useUser } from '../../Context/useUser';
+import { useUser } from '../../Context/User/useUser';
 import { CapitaliseAfterSpace } from '../../utils/CapitaliseAfterSpace';
 import EditBoardApi from '../../packages/Api/BoardApi/EditBoardApi';
 import { openCustomNotification } from '../../utils/notificationUtil';
