+++ frontend/src/components/Containers/EditDeleteContainer.tsx
@@ -5,7 +5,7 @@ import {
   onClickPropsForDelete,
 } from '../Interface/AddTaskInterface';
 import './ContainersStyles.css';
-import { useTheme } from '../../Context/UseTheme';
+import { useTheme } from '../../Context/Theme/UseTheme';
 import { motion, AnimatePresence } from 'framer-motion';
